requirements.txt
selenium==4.16.0
webdriver-manager==4.0.1
pandas==2.1.4
openpyxl==3.1.2
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.0
